Move lexical-editor under app/edit/elements
