+++ components/form-builder/app/edit/elements/RichText.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useTemplateStore } from "../../../store/useTemplateStore";
-import { RichTextEditor } from "../../../lexical-editor/RichTextEditor";
+import { RichTextEditor } from "./lexical-editor/RichTextEditor";
 import { LocalizedElementProperties } from "../../../types";
 import { useTranslation } from "next-i18next";
