+++ components/form-builder/app/edit/elements/lexical-editor/config.ts
