+++ components/form-builder/app/edit/elements/lexical-editor/plugins/FocusEditor.tsx
