+++ components/form-builder/app/edit/elements/lexical-editor/utils/getSelectedNode.ts
