+++ components/form-builder/app/edit/elements/lexical-editor/utils/setFloatingElemPosition.ts
